refactor(compare): unify comparison artifact generation for CLI and xtask (#1309)
Move the canonical write_comparison_artifacts writer into provenant::compare
with file-ownership support, and make compare-outputs call it via a thin
wrapper. Sole trailing hex ?uuid= is scan-local by PackageUid construction.
